+++ src/scripts/contract.read.ts
