+++ tsconfig.json
@@ -63,5 +63,5 @@
         "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */,
         "resolveJsonModule": true
     },
-    "include": ["sources/**/*"]
+    "include": ["src/**/*"]
 }
